backend/requirements.txt
fastapi
uvicorn
python-multipart
fastai
torch
torchvision
pillow
